增加数据库迁移回滚锁开关，防止误删除

+++ database/migrations/2020_12_04_152340_sys_node.php
@@ -39,6 +39,8 @@ class SysNode extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists($this->tablename);
+        if (true === COMM_DBLOCK) {
+            Schema::dropIfExists($this->tablename);
+        }
     }
 }
