+++ database/migrations/2020_12_04_152351_sys_user.php
@@ -59,6 +59,8 @@ class SysUser extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists($this->tablename);
+        if (true === COMM_DBLOCK) {
+            Schema::dropIfExists($this->tablename);
+        }
     }
 }
