+++ database/migrations/2020_12_05_185035_log_userlogin.php
@@ -39,6 +39,8 @@ class LogUserlogin extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists($this->tablename);
+        if (true === COMM_DBLOCK) {
+            Schema::dropIfExists($this->tablename);
+        }
     }
 }
