+++ database/migrations/2020_12_21_171802_tpl_example.php
@@ -52,6 +52,8 @@ class TplExample extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists($this->tablename);
+        if (true === COMM_DBLOCK) {
+            Schema::dropIfExists($this->tablename);
+        }
     }
 }
